requirements.txt
asgiref==3.9.1
cffi==1.17.1
dj-database-url==0.5.0
Django==4.2.23
django-allauth==65.11.1
django-tailwind==4.2.0
gunicorn==20.1.0
numpy==2.3.2
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.1.1
scipy==1.16.1
setuptools==80.9.0
soundfile==0.13.1
sqlparse==0.5.3
whitenoise==5.3.0
